requirements.txt
pillow
streamlit
numpy
pandas
requests